Fully enable ComWrappers for COM TypeDescriptor (#9278)
Uses ComHelpers to query for interface pointers to enable all COM scenarios (ComWrappers, Com Interop, our own Com exposed classes).

- Change Com2DataTypeToManagedDataTypeConverter to go between VARIANT and object.
- Remove Activator.CreateInstance call for type converters from Com2PropertyDescriptor and just create the instances directly as needed.
- Add a NullTypeDescriptor to simplify the ComTypeDescriptor.
- Add reverse IUnknown* cast operator to VARIANT.
- Ensure we don't return null from BSTR.ToStringAndFree and only free if it isn't null.
- Change RegisterInterfaceInGlobal to return HRESULT for easier debugging.
- Use centralized IDispatch property gettter/setter helpers and ensure we free any returned exception BSTRs.
- Add IVTable to additional COM structs.
- Add tests to cover both legacy COM and ComWrappers and also validate IProvideMulitpleClassInfo
- Move IDispatchTests to the right namespace.
- Seal classes, revise variable names, split out COM struct files, and other various cleanup.

+++ src/System.Windows.Forms/src/System/Windows/Forms/Control.ActiveXImpl.cs
@@ -611,7 +611,7 @@ public partial class Control
             Debug.Indent();
             Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "clientSite implements IDispatch");
 
-            hr = dispatch.Value->TryGetProperty((uint)dispid, &property, PInvoke.LCID.USER_DEFAULT.RawValue);
+            hr = dispatch.Value->TryGetProperty(dispid, &property, PInvoke.LCID.USER_DEFAULT.RawValue);
 
             if (hr.Succeeded)
             {
